🏗️ Dimension measurements based on another.
I realize that it is shaky, since if you give it a
value that matches the type description but isn't
based on the same prototypal black magic, it will
explode. Please open a ticket if you actually face
a bug because of this and I'm more than happy to
address it properly.

Please do understand that I'm wrestling with this
lib's best contract, abstraction and signatures for
two weeks non-stop, day and night, literally,
therefore at this point I'm too lazy to cover for
the edge cases that can of-course occur in
production.

Again, you'll only have problems if you produce
a dimension measurement that is not based on the
function provided by this library! Which is supposed
to be supported by design.